DOCS: Added a simple plan, which needs shelling out.
This is mostly to get the ball rolling, the jokes need
to be reduced down to actual performable tasks, and the
language might need a bit of working. Currently the
plan exists only in nynorsk, mainly to reflect the
choice of Java as the language of choice for the project.

Any takers for making a proper plan?